+++ delivery/src/main/java/com/faster/delivery/app/delivery/application/dto/DeliverySaveDto.java
@@ -1,0 +1,12 @@
+package com.faster.delivery.app.delivery.application.dto;
+
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record DeliverySaveDto(UUID orderId,
+                              UUID sourceHubId,
+                              UUID destinationHubId,
+                              UUID receiveCompanyId) {
+
+}
