+++ delivery/src/main/java/com/faster/delivery/app/delivery/application/usecase/DeliveryService.java
@@ -1,0 +1,8 @@
+package com.faster.delivery.app.delivery.application.usecase;
+
+import com.faster.delivery.app.delivery.application.dto.DeliverySaveDto;
+import java.util.UUID;
+
+public interface DeliveryService {
+  UUID saveDelivery(Long userId, DeliverySaveDto deliverySaveDto);
+}
