+++ delivery/src/main/java/com/faster/delivery/app/delivery/domain/repository/DeliveryRepository.java
@@ -1,0 +1,7 @@
+package com.faster.delivery.app.delivery.domain.repository;
+
+import com.faster.delivery.app.delivery.domain.entity.Delivery;
+
+public interface DeliveryRepository {
+  Delivery save(Delivery delivery);
+}
