+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/feign/HubClient.java
@@ -1,0 +1,14 @@
+package com.faster.delivery.app.delivery.infrastructure.feign;
+
+import com.common.response.ApiResponse;
+import com.faster.delivery.app.delivery.infrastructure.feign.dto.hub.HubPathRequestDto;
+import com.faster.delivery.app.delivery.infrastructure.feign.dto.hub.HubPathResponseDto;
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+
+@FeignClient(name = "hub-service", contextId = "hubClientForDelivery")
+public interface HubClient {
+  @PostMapping("/internal/hubs/paths")
+  ApiResponse<HubPathResponseDto> getHubRouteData(@RequestBody HubPathRequestDto hubPathRequestDto);
+}
