+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/jpa/DeliveryRepositoryCustom.java
@@ -1,0 +1,5 @@
+package com.faster.delivery.app.delivery.infrastructure.jpa;
+
+public interface DeliveryRepositoryCustom {
+
+}
