+++ delivery/src/main/java/com/faster/delivery/app/delivery/infrastructure/jpa/DeliveryRepositoryCustomImpl.java
@@ -1,0 +1,5 @@
+package com.faster.delivery.app.delivery.infrastructure.jpa;
+
+public class DeliveryRepositoryCustomImpl implements DeliveryRepositoryCustom {
+
+}
