+++ delivery/src/main/java/com/faster/delivery/app/deliverymanager/application/HubClient.java
@@ -1,0 +1,8 @@
+package com.faster.delivery.app.deliverymanager.application;
+
+import com.faster.delivery.app.deliverymanager.application.dto.HubDto;
+import java.util.UUID;
+
+public interface HubClient {
+  HubDto getHubData(UUID hubId);
+}
