+++ delivery/src/main/java/com/faster/delivery/app/deliverymanager/application/UserClient.java
@@ -1,0 +1,7 @@
+package com.faster.delivery.app.deliverymanager.application;
+
+import com.faster.delivery.app.deliverymanager.application.dto.UserDto;
+
+public interface UserClient {
+  UserDto getUserData(Long userId);
+}
