+++ delivery/src/main/java/com/faster/delivery/app/deliverymanager/application/dto/DeliveryManagerSaveDto.java
@@ -1,0 +1,12 @@
+package com.faster.delivery.app.deliverymanager.application.dto;
+
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record DeliveryManagerSaveDto(
+    UUID hubId,
+    String type,
+    Long userId
+) {
+}
