+++ delivery/src/main/java/com/faster/delivery/app/deliverymanager/domain/repository/DeliveryManagerRepository.java
@@ -1,0 +1,7 @@
+package com.faster.delivery.app.deliverymanager.domain.repository;
+
+import com.faster.delivery.app.deliverymanager.domain.entity.DeliveryManager;
+
+public interface DeliveryManagerRepository {
+  DeliveryManager save(DeliveryManager deliveryManager);
+}
