+++ delivery/src/main/java/com/faster/delivery/app/deliverymanager/infrastructure/jpa/DeliveryManagerJpaRepository.java
@@ -1,0 +1,10 @@
+package com.faster.delivery.app.deliverymanager.infrastructure.jpa;
+
+import com.faster.delivery.app.deliverymanager.domain.entity.DeliveryManager;
+import com.faster.delivery.app.deliverymanager.domain.repository.DeliveryManagerRepository;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DeliveryManagerJpaRepository extends DeliveryManagerRepository, JpaRepository<DeliveryManager, UUID>, DeliveryManagerRepositoryCustom {
+
+}
