+++ delivery/src/main/java/com/faster/delivery/app/deliverymanager/infrastructure/jpa/DeliveryManagerRepositoryCustomImpl.java
@@ -1,0 +1,5 @@
+package com.faster.delivery.app.deliverymanager.infrastructure.jpa;
+
+public class DeliveryManagerRepositoryCustomImpl implements DeliveryManagerRepositoryCustom {
+
+}
