+++ delivery/src/main/java/com/faster/delivery/app/deliverymanager/presentation/dto/DeliveryManagerSaveRequestDto.java
@@ -1,0 +1,21 @@
+package com.faster.delivery.app.deliverymanager.presentation.dto;
+
+import com.faster.delivery.app.deliverymanager.application.dto.DeliveryManagerSaveDto;
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record DeliveryManagerSaveRequestDto(
+  UUID hubId,
+  String type,
+  Long userId
+) {
+
+  public DeliveryManagerSaveDto toSaveDto() {
+    return DeliveryManagerSaveDto.builder()
+        .hubId(this.hubId)
+        .userId(this.userId)
+        .type(this.type)
+        .build();
+  }
+}
