--- delivery/src/main/java/com/faster/delivery/app/domains/delivery/domain/repository/DeliveryRepository.java
@@ -1,7 +1,0 @@
-package com.faster.delivery.app.domains.delivery.domain.repository;
-
-import com.faster.delivery.app.domains.delivery.domain.entity.Delivery;
-
-public interface DeliveryRepository {
-  Delivery save(Delivery delivery);
-}
