--- delivery/src/main/java/com/faster/delivery/app/domains/delivery/infrastructure/jpa/DeliveryRepositoryCustom.java
@@ -1,5 +1,0 @@
-package com.faster.delivery.app.domains.delivery.infrastructure.jpa;
-
-public interface DeliveryRepositoryCustom {
-
-}
