--- delivery/src/main/java/com/faster/delivery/app/domains/delivery/infrastructure/jpa/DeliveryRepositoryCustomImpl.java
@@ -1,5 +1,0 @@
-package com.faster.delivery.app.domains.delivery.infrastructure.jpa;
-
-public class DeliveryRepositoryCustomImpl implements DeliveryRepositoryCustom {
-
-}
